refactor(layout): chop it down into multiple chunks and add the footer
