--- components/layout.tsx
@@ -1,52 +1,0 @@
-import { Page, Text, useMediaQuery } from '@geist-ui/core'
-import Head from 'next/head'
-import { FC, ReactElement } from 'react'
-import { useSectionWidthMediaQuery } from './hooks/useSectionWidthMediaQuery'
-
-export const SITE_NAME = 'Flatter'
-
-const Layout: FC<{ children: ReactElement; header: ReactElement }> = ({
-    children,
-    header,
-}) => {
-    const width = useSectionWidthMediaQuery()
-    return (
-        <Page dotBackdrop width={width}>
-            <Head>
-                <link rel="icon" href="/favicon.svg" />
-                <meta
-                    name="description"
-                    content="Search the latest and greatest property deals for Berlin"
-                />
-                <meta
-                    property="og:image"
-                    content={`https://og-image.vercel.app/${encodeURI(
-                        SITE_NAME
-                    )}.png?theme=light&md=0&fontSize=75px&images=https%3A%2F%2Fassets.zeit.co%2Fimage%2Fupload%2Ffront%2Fassets%2Fdesign%2Fnextjs-black-logo.svg`}
-                />
-                <meta name="og:title" content={SITE_NAME} />
-                <meta property="og:type" content="website" />
-                <meta property="og:url" content="https://flat.marpme.fyi" />
-                <meta name="twitter:card" content="summary_large_image" />
-                <title>{SITE_NAME} ~ Berlin</title>
-            </Head>
-            <header
-                style={{
-                    marginBottom: '2em',
-                }}
-            >
-                <Text
-                    h1
-                    style={{ letterSpacing: '.03em' }}
-                    data-testid="header-name"
-                >
-                    🏘️ {SITE_NAME}
-                </Text>
-                {header}
-            </header>
-            {children}
-        </Page>
-    )
-}
-
-export default Layout
